pyproject.toml
[project]
name = "nidus"
version = "0.1.0"
description = "Nidus: open-source document store database"
readme = "README.md"
requires-python = "==3.14.*"
dependencies = [
    "mcp-server",
    "cli",
    "common",
]

[tool.uv]
package = true
constraint-dependencies = [
    "pydantic-settings>=2.13.1",
    "lancedb>=0.29.2",
    "pyarrow>=23.0.1",
    "model2vec>=0.7.0",
    "huggingface-hub>=1.4.1",
    "mcp>=1.26.0",
    "mistune>=3.2.0",
    "pypdf>=6.7.1",
    "pdfminer-six>=20260107",
    "click>=8.3.1",
    "watchdog>=6.0.0",
]

[tool.uv.workspace]
members = ["packages/*"]

[dependency-groups]
dev = [
    "pip-licenses>=5.5.1",
    "pyright[nodejs]>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]

[tool.pytest.ini_options]
markers = [
    "small: no I/O, no network, single process",
    "medium: local file or LanceDB access allowed, no external network",
    "large: external services and network allowed",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "RUF", # ruff-specific
]
ignore = [
    "RUF001", # 日本語全角文字（文字列）を許容
    "RUF002", # 日本語全角文字（docstring）を許容
    "RUF003", # 日本語全角文字（コメント）を許容
]

[tool.setuptools.packages.find]
include = ["packages*"]

[project.scripts]
nidus = "cli.main:main"
nidus-mcp = "mcp_server.main:main"

[tool.uv.sources]
mcp-server = { workspace = true }
common = { workspace = true }
cli = { workspace = true }
